move credential store class
